% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vassilakis.R
\name{dyad_roughness_vass}
\alias{dyad_roughness_vass}
\title{Dyad roughness (Vassilakis)}
\usage{
dyad_roughness_vass(f1, f2, a1, a2)
}
\arguments{
\item{f1}{Frequency of tone 1 (Hz) (numeric vector)}

\item{f2}{Frequency of tone 2 (Hz) (numeric vector)}

\item{a1}{amplitude of tone 1 (numeric vector)}

\item{a2}{amplitude of tone 2 (numeric vector)}
}
\value{
Numeric vector of roughnesses
\insertRef{Vassilakis2001}{dycon}

\insertRef{Villegas2010}{dycon}
}
\description{
Gets the roughness of a dyad according to the model of Vassilakis (2001) as reported in Villegas (2010).
}
