% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sethares.R
\name{get_roughness_seth}
\alias{get_roughness_seth}
\title{Spectral roughness (Sethares)}
\usage{
get_roughness_seth(frequency, amplitude, seth_params = get_seth_params())
}
\arguments{
\item{frequency}{Numeric vector of frequencies}

\item{amplitude}{Numeric vector of amplitudes}
}
\value{
Numeric vector of roughnesses
\insertRef{Sethares1993}{dycon}
}
\description{
Gets the roughness of a spectrum according to the model of Sethares (1993).
}
