% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vassilakis.R
\name{get_roughness_vass}
\alias{get_roughness_vass}
\title{Spectral roughness (Vassilakis)}
\usage{
get_roughness_vass(frequency, amplitude)
}
\arguments{
\item{frequency}{Numeric vector of frequencies}

\item{amplitude}{Numeric vector of amplitudes}
}
\value{
Numeric vector of roughnesses
\insertRef{Vassilakis2001}{dycon}

\insertRef{Villegas2010}{dycon}
}
\description{
Gets the roughness of a spectrum according to the model of Vassilakis (2001) as reported in Villegas (2010).
}
