% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutch.R
\docType{data}
\name{hutch_visualise_theme}
\alias{hutch_visualise_theme}
\title{ggplot theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
hutch_visualise_theme
}
\description{
Defines a default theme for visualising computations
for Hutchinson & Knopoff's (1978) model
(see \code{\link{hutch_visualise}}).
}
\keyword{datasets}
