% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutch-knopoff.R
\name{roughness_hutch}
\alias{roughness_hutch}
\title{Spectral roughness (Hutchinson & Knopoff)}
\usage{
roughness_hutch(frequency, amplitude, cbw_cut_off = 1.2, a = 0.25,
  b = 2)
}
\arguments{
\item{frequency}{Numeric vector of frequencies.}

\item{amplitude}{Numeric vector of amplitudes.}

\item{cbw_cut_off}{Parameter passed to \code{\link{hutch_g}()}.}

\item{a}{Parameter passed to \code{\link{hutch_g}()}.}

\item{b}{Parameter passed to \code{\link{hutch_g}()}.}
}
\value{
Numeric scalar, identifying the roughness of the spectrum.
}
\description{
Gets the roughness of a spectrum according to the model of
\insertCite{Hutchinson1978;textual}{dycon}.
}
\references{
\insertAllCited{}
}
