% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sethares.R
\name{roughness_seth}
\alias{roughness_seth}
\title{Spectral roughness (Sethares)}
\usage{
roughness_seth(frequency, amplitude, ...)
}
\arguments{
\item{frequency}{Numeric vector of frequencies.}

\item{amplitude}{Numeric vector of amplitudes.}

\item{...}{Further parameters to be passed to \code{\link{dyad_roughness_seth}}.}
}
\value{
Numeric vector of roughnesses.
}
\description{
Gets the roughness of a spectrum according to the model of Sethares (1993).
By default, the algorithm is modified according to
\insertCite{Sethares2005;textual}{dycon} and
\insertCite{Weisser2013;textual}{dycon}:
roughness is proportional to the minimum amplitude of each pair of partials,
not the product of their amplitudes.
This behaviour can be disabled by setting \code{min_amplitude = FALSE}.
}
\note{
The function assumes that any input complex tones have already
been expanded into their constituent pure tones.

\insertCite{Sethares2005;textual}{dycon}
suggests using loudnesses instead of amplitudes.
However, he acknowledges that loudness is difficult to calculate
for arbitrary timbres.
Furthermore, if we replace amplitude with roughness,
we lose the original model's invariance to multiplicative
scaling of the original signal.
In this implementation, we therefore stay with amplitude,
consistent with \insertCite{Sethares1993;textual}{dycon}.
}
\references{
\insertAllCited{}
}
