% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vassilakis.R
\name{roughness_vass}
\alias{roughness_vass}
\alias{roughness_vass.default}
\alias{roughness_vass.fr_sparse_spectrum}
\title{Spectral roughness (Vassilakis)}
\usage{
roughness_vass(x, ...)

\method{roughness_vass}{default}(x, ...)

\method{roughness_vass}{fr_sparse_spectrum}(x)
}
\arguments{
\item{x}{Object to analyse, which is coerced to the class
\code{\link[hrep]{fr_sparse_spectrum}}.
If the input is numeric, it will be treated as a vector of MIDI note numbers,
and expanded to its implied harmonic spectrum
(see \code{\link[hrep]{fr_sparse_spectrum}}).}

\item{...}{Further arguments to pass to \code{\link[hrep]{fr_sparse_spectrum}}.}
}
\value{
Estimated roughness, as a numeric scalar.
}
\description{
Gets the roughness of a spectrum according to the model of
\insertCite{Vassilakis2001;textual}{dycon}
\insertCite{Villegas2010;textual}{dycon}
}
\references{
\insertAllCited{}
}
