% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutch.R
\name{roughness_hutch}
\alias{roughness_hutch}
\alias{roughness_hutch.default}
\alias{roughness_hutch.fr_sparse_spectrum}
\title{Spectral roughness (Hutchinson & Knopoff)}
\usage{
roughness_hutch(x, cbw_cut_off = 1.2, a = 0.25, b = 2, ...)

\method{roughness_hutch}{default}(x, cbw_cut_off = 1.2, a = 0.25,
  b = 2, ...)

\method{roughness_hutch}{fr_sparse_spectrum}(x, cbw_cut_off = 1.2,
  a = 0.25, b = 2)
}
\arguments{
\item{x}{Object to analyse, which is coerced to the class
\code{\link[hrep]{fr_sparse_spectrum}}.
If the input is numeric, it will be treated as a vector of MIDI note numbers,
and expanded to its implied harmonic spectrum
(see \code{\link[hrep]{fr_sparse_spectrum}}).}

\item{cbw_cut_off}{Parameter passed to \code{\link{hutch_g}()}.}

\item{a}{Parameter passed to \code{\link{hutch_g}()}.}

\item{b}{Parameter passed to \code{\link{hutch_g}()}.}

\item{...}{Further arguments to pass to \code{\link[hrep]{fr_sparse_spectrum}}.}
}
\value{
Numeric scalar, identifying the roughness of the spectrum.
}
\description{
Gets the roughness of a sonority according to the model of
\insertCite{Hutchinson1978;textual}{dycon}.
}
\references{
\insertAllCited{}
}
