% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vassilakis.R
\name{roughness_vass}
\alias{roughness_vass}
\title{Spectral roughness (Vassilakis)}
\usage{
roughness_vass(frequency, amplitude)
}
\arguments{
\item{frequency}{Numeric vector of frequencies.}

\item{amplitude}{Numeric vector of amplitudes.}
}
\value{
Numeric vector of roughnesses.
}
\description{
Gets the roughness of a spectrum according to the model of
\insertCite{Vassilakis2001;textual}{dycon}
\insertCite{Villegas2010;textual}{dycon}
}
\note{
The function assumes that any input complex tones have already
been expanded into their constituent pure tones.
}
\references{
\insertAllCited{}
}
